polygl0ts-friday-meetings/Bad_Mitigations/requirements.txt
flask
dnspython
requests